Updated OpenAPI specification
Updated the OpenAPI specification to now include a path for receiving completed Tasks on the webservice, as well as defining a schema for the structure of a completed Task.
Fixed an existing line in the existing OpenAPI specification which used syntax which are not supported in the defined version of OpenAPI specification.